requirements_scraping.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
psycopg2-binary>=2.9.7
